Add JWT error handling and admin claim checks to item routes

+++ resources/item.py
@@ -2,6 +2,7 @@ import uuid
 
 from flask import request
 from flask.views import MethodView
+from flask_jwt_extended import jwt_required, get_jwt
 from flask_smorest import Blueprint, abort  # type: ignore
 from sqlalchemy.exc import SQLAlchemyError
 
@@ -15,6 +16,7 @@ blp = Blueprint("Items", __name__, description="Operations on ITEMS.")
 # registers this Item class as the handler for /item/<item_id>
 @blp.route("/item/<string:item_id>")
 class Item(MethodView):  # contains all HTTP methods (mapped to the methods) for /item/<item_id>
+    @jwt_required()  # jwt token required to access this route and method
     @blp.response(200, ItemSchema)  # updates the docs, returns status code
     # Serialize the response using ItemSchema and return HTTP 200 OK to the API client
     def get(self, item_id):
@@ -22,7 +24,17 @@ class Item(MethodView):  # contains all HTTP methods (mapped to the methods) for
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required()
     def delete(self, item_id):
+
+        # using the JWT claims
+        # Retrieve the entire decoded JWT payload (claims) from the current request.
+        jwt = get_jwt()
+        # Check for a custom claim 'is_admin' in the JWT.
+        # This claim would have been added when the token was created, indicating user privileges.
+        if not jwt.get("is_admin"):
+            abort(401, exc="Admin privilege required.")
+
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -61,12 +73,14 @@ class Item(MethodView):  # contains all HTTP methods (mapped to the methods) for
 
 @blp.route("/item")
 class ItemList(MethodView):
+    @jwt_required()
     # many=True indicates that the response will be a list of items!
     @blp.response(200, ItemSchema(many=True))
     # Handles GET requests to /item
     def get(self):
         return ItemModel.query.all()  # returns all the records in the item table
 
+    @jwt_required()
     # Validates incoming JSON against ItemSchema
     # If validation passes, the deserialized data is injected as `item_data`
     # If validation fails, Flask-Smorest automatically returns a 400 Bad Request with an error message
